test(egress): reuse existing Linux regression test file
Keep the timed-element regression unchanged while reusing the existing licensed Linux test file. This avoids an extra source file and satisfies repository license checks.
